+++ crates/maple_core/src/dirs.rs
@@ -1,0 +1,13 @@
+use directories::{BaseDirs, ProjectDirs};
+use once_cell::sync::Lazy;
+
+/// Project directory for Vim Clap.
+///
+/// All the files generated by vim-clap are stored there.
+pub static PROJECT_DIRS: Lazy<ProjectDirs> = Lazy::new(|| {
+    ProjectDirs::from("org", "vim", "Vim Clap")
+        .expect("Couldn't create project directory for vim-clap")
+});
+
+pub static BASE_DIRS: Lazy<BaseDirs> =
+    Lazy::new(|| BaseDirs::new().expect("Failed to construct BaseDirs"));
