+++ crates/maple_core/src/find_usages/mod.rs
@@ -1,9 +1,68 @@
 mod search_engine;
 
+use matcher::{ExactMatcher, InverseMatcher};
 use rayon::prelude::*;
 use std::ops::{Index, IndexMut};
+use types::{CaseMatching, ExactTerm, InverseTerm};
 
 pub use self::search_engine::{CtagsSearcher, GtagsSearcher, QueryType, RegexSearcher};
+
+/// Matcher for filtering out the unqualified usages earlier at the searching stage.
+#[derive(Debug, Clone, Default)]
+pub struct UsageMatcher {
+    pub exact_matcher: ExactMatcher,
+    pub inverse_matcher: InverseMatcher,
+}
+
+impl UsageMatcher {
+    pub fn new(exact_terms: Vec<ExactTerm>, inverse_terms: Vec<InverseTerm>) -> Self {
+        Self {
+            exact_matcher: ExactMatcher::new(exact_terms, CaseMatching::Smart),
+            inverse_matcher: InverseMatcher::new(inverse_terms),
+        }
+    }
+
+    /// Returns the match indices of exact terms if given `line` passes all the checks.
+    fn match_indices(&self, line: &str) -> Option<Vec<usize>> {
+        match (
+            self.exact_matcher.find_matches(line),
+            self.inverse_matcher.match_any(line),
+        ) {
+            (Some((_, indices)), false) => Some(indices),
+            _ => None,
+        }
+    }
+
+    /// Returns `true` if the result of The results of applying `self`
+    /// is a superset of applying `other` on the same source.
+    pub fn is_superset(&self, other: &Self) -> bool {
+        self.exact_matcher
+            .exact_terms()
+            .iter()
+            .zip(other.exact_matcher.exact_terms().iter())
+            .all(|(local, other)| local.is_superset(other))
+            && self
+                .inverse_matcher
+                .inverse_terms()
+                .iter()
+                .zip(other.inverse_matcher.inverse_terms().iter())
+                .all(|(local, other)| local.is_superset(other))
+    }
+
+    pub fn match_jump_line(
+        &self,
+        (jump_line, mut indices): (String, Vec<usize>),
+    ) -> Option<(String, Vec<usize>)> {
+        if let Some(exact_indices) = self.match_indices(&jump_line) {
+            indices.extend_from_slice(&exact_indices);
+            indices.sort_unstable();
+            indices.dedup();
+            Some((jump_line, indices))
+        } else {
+            None
+        }
+    }
+}
 
 #[derive(Clone, Debug, Default)]
 pub struct Usage {
