+++ crates/maple_core/src/lib.rs
@@ -1,6 +1,7 @@
 mod cache;
 pub mod config;
 pub mod datastore;
+pub mod dirs;
 pub mod find_usages;
 pub mod helptags;
 pub mod paths;
