+++ crates/maple_core/src/utils.rs
@@ -1,84 +1,12 @@
+use crate::dirs::{BASE_DIRS, PROJECT_DIRS};
 use anyhow::Result;
 use chrono::prelude::*;
-use directories::ProjectDirs;
 use icon::Icon;
-use itertools::Itertools;
-use matcher::{ExactMatcher, InverseMatcher};
 use once_cell::sync::Lazy;
-use std::borrow::Cow;
 use std::io::{BufRead, BufReader, Lines};
 use std::path::{Path, PathBuf};
 use subprocess::Exec;
-use types::{CaseMatching, ExactTerm, InverseTerm};
 use utility::{println_json, println_json_with_length, read_first_lines};
-
-/// Project directory for Vim Clap.
-///
-/// All the files generated by vim-clap are stored there.
-pub static PROJECT_DIRS: Lazy<ProjectDirs> = Lazy::new(|| {
-    ProjectDirs::from("org", "vim", "Vim Clap")
-        .expect("Couldn't create project directory for vim-clap")
-});
-
-pub static BASE_DIRS: Lazy<directories::BaseDirs> =
-    Lazy::new(|| directories::BaseDirs::new().expect("Failed to construct BaseDirs"));
-
-/// Yes or no terms.
-#[derive(Debug, Clone, Default)]
-pub struct UsageMatcher {
-    pub exact_matcher: ExactMatcher,
-    pub inverse_matcher: InverseMatcher,
-}
-
-impl UsageMatcher {
-    pub fn new(exact_terms: Vec<ExactTerm>, inverse_terms: Vec<InverseTerm>) -> Self {
-        Self {
-            exact_matcher: ExactMatcher::new(exact_terms, CaseMatching::Smart),
-            inverse_matcher: InverseMatcher::new(inverse_terms),
-        }
-    }
-
-    /// Returns the match indices of exact terms if given `line` passes all the checks.
-    fn match_indices(&self, line: &str) -> Option<Vec<usize>> {
-        match (
-            self.exact_matcher.find_matches(line),
-            self.inverse_matcher.match_any(line),
-        ) {
-            (Some((_, indices)), false) => Some(indices),
-            _ => None,
-        }
-    }
-
-    /// Returns `true` if the result of The results of applying `self`
-    /// is a superset of applying `other` on the same source.
-    pub fn is_superset(&self, other: &Self) -> bool {
-        self.exact_matcher
-            .exact_terms()
-            .iter()
-            .zip(other.exact_matcher.exact_terms().iter())
-            .all(|(local, other)| local.is_superset(other))
-            && self
-                .inverse_matcher
-                .inverse_terms()
-                .iter()
-                .zip(other.inverse_matcher.inverse_terms().iter())
-                .all(|(local, other)| local.is_superset(other))
-    }
-
-    pub fn match_jump_line(
-        &self,
-        (jump_line, mut indices): (String, Vec<usize>),
-    ) -> Option<(String, Vec<usize>)> {
-        if let Some(exact_indices) = self.match_indices(&jump_line) {
-            indices.extend_from_slice(&exact_indices);
-            indices.sort_unstable();
-            indices.dedup();
-            Some((jump_line, indices))
-        } else {
-            None
-        }
-    }
-}
 
 pub type UtcTime = DateTime<Utc>;
 
@@ -180,13 +108,6 @@ pub(crate) fn expand_tilde(path: impl AsRef<str>) -> PathBuf {
     }
 }
 
-/// Build the absolute path using cwd and relative path.
-pub fn build_abs_path(cwd: impl AsRef<Path>, curline: impl AsRef<Path>) -> PathBuf {
-    let mut path: PathBuf = cwd.as_ref().into();
-    path.push(curline);
-    path
-}
-
 /// Counts lines in the source `handle`.
 ///
 /// # Examples
@@ -237,66 +158,6 @@ pub fn display_width(n: usize) -> usize {
     len
 }
 
-// /home/xlc/.rustup/toolchains/stable-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/alloc/src/string.rs
-pub(crate) fn truncate_absolute_path(abs_path: &str, max_len: usize) -> Cow<'_, str> {
-    if abs_path.len() > max_len {
-        let gap = abs_path.len() - max_len;
-
-        const SEP: char = std::path::MAIN_SEPARATOR;
-
-        if let Some(home_dir) = BASE_DIRS.home_dir().to_str() {
-            if abs_path.starts_with(home_dir) {
-                // ~/.rustup/toolchains/stable-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/alloc/src/string.rs
-                if home_dir.len() > gap {
-                    return abs_path.replacen(home_dir, "~", 1).into();
-                }
-
-                // ~/.rustup/.../github.com/paritytech/substrate/frame/system/src/lib.rs
-                let home_stripped = &abs_path.trim_start_matches(home_dir)[1..];
-                if let Some((first, target)) = home_stripped.split_once(SEP) {
-                    let mut hidden = 0usize;
-                    for component in target.split(SEP) {
-                        if hidden > gap + 2 {
-                            let mut target = target.to_string();
-                            target.replace_range(..hidden - 1, "...");
-                            return format!("~{SEP}{first}{SEP}{target}").into();
-                        } else {
-                            hidden += component.len() + 1;
-                        }
-                    }
-                }
-            } else {
-                let top = abs_path.splitn(4, SEP).collect::<Vec<_>>();
-                if let Some(last) = top.last() {
-                    if let Some((_first, target)) = last.split_once(SEP) {
-                        let mut hidden = 0usize;
-                        for component in target.split(SEP) {
-                            if hidden > gap + 2 {
-                                let mut target = target.to_string();
-                                target.replace_range(..hidden - 1, "...");
-                                let head = top.iter().take(top.len() - 1).join(&SEP.to_string());
-                                return format!("{head}{SEP}{target}").into();
-                            } else {
-                                hidden += component.len() + 1;
-                            }
-                        }
-                    }
-                }
-            }
-        } else {
-            // Truncate the left of absolute path string.
-            // ../stable-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/alloc/src/string.rs
-            if let Some((offset, _)) = abs_path.char_indices().nth(abs_path.len() - max_len + 2) {
-                let mut abs_path = abs_path.to_string();
-                abs_path.replace_range(..offset, "..");
-                return abs_path.into();
-            }
-        }
-    }
-
-    abs_path.into()
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -305,29 +166,5 @@ mod tests {
     fn test_count_lines() {
         let f: &[u8] = b"some text\nwith\nfour\nlines\n";
         assert_eq!(count_lines(f).unwrap(), 4);
-    }
-
-    #[test]
-    fn test_truncate_absolute_path() {
-        #[cfg(not(target_os = "windows"))]
-        let p = ".rustup/toolchains/stable-x86_64-unknown-linux-gnu/lib/rustlib/src/rust/library/alloc/src/string.rs";
-        #[cfg(target_os = "windows")]
-        let p = r#".rustup\toolchains\stable-x86_64-unknown-linux-gnu\lib\rustlib\src\rust\library\alloc\src\string.rs"#;
-        let abs_path = format!(
-            "{}{}{}",
-            BASE_DIRS.home_dir().to_str().unwrap(),
-            std::path::MAIN_SEPARATOR,
-            p
-        );
-        let max_len = 60;
-        #[cfg(not(target_os = "windows"))]
-        let expected = "~/.rustup/.../src/rust/library/alloc/src/string.rs";
-        #[cfg(target_os = "windows")]
-        let expected = r#"~\.rustup\...\src\rust\library\alloc\src\string.rs"#;
-        assert_eq!(truncate_absolute_path(&abs_path, max_len), expected);
-
-        let abs_path = "/media/xlc/Data/src/github.com/paritytech/substrate/bin/node/cli/src/command_helper.rs";
-        let expected = "/media/xlc/.../bin/node/cli/src/command_helper.rs";
-        assert_eq!(truncate_absolute_path(abs_path, max_len), expected);
     }
 }
